* CalculatorLibrary.csproj: #2 - Create a Class Library for the   Calculator Application Assignment
* finalProject.sln:
* Class1.cs:
* SimpleCalc.csproj:
* SimpleCalc.cs:
* finalProject.csproj:
* mono_crash.mem.29223.1.blob:
* CalculatorController.cs:

* ExpressionModel.cs: #2 - Create a Class Library for the Calculator
  Application Assignment

+++ finalProject/Models/ExpressionModel.cs
@@ -3,9 +3,9 @@ namespace finalProject.Models
 {
     public class ExpressionModel
     {
-        public double operand1 { get; set; }
+        public decimal operand1 { get; set; }
 
-        public double operand2 { get; set; }
+        public decimal operand2 { get; set; }
 
         public string operator_symbol { get; set; }  //named operator_symbol as operator seems to be keyword
